Revert circuit variation strategy
The circuits that each client creates was varied accross clients
as shown in this commit. This is the strategy that was used in
the "Firefox Scale" paper.

We thought that removing the uniform distribution mod from 0.25 to
1.75 was no longer necessary since we use randomly seeded flow
Markov models. However, an initial experiment after removing the
uniform distribution mod has shown a significant increase in the
number of tgen PROXY-STALLOUT-CONNECT errors.

Until we can debug and correct the underlying issue, we are for
now reverting to the strategy that has been heavily tested and
known to produce good results.