More flexibility for running simulations
This allows us to fully customize the argument string that we pass
to shadow when running a simulation. This is particularly useful
when running simulations with v1.15.x, since the args that were
hard-coded here before were not available in v1.15.0.

Also, we now try to compress both yaml and xml config files if
they exist.