parse_onionperf: use helper to get recv timestamps
Unify more of the goodput calculation logic by using a helper that
dynamically looks for a given timestamp in the stream's
payload_bytes_recv, and then falls back to payload_progress_recv. This
makes it "transparent" that payload_bytes_recv doesn't have a 4 MB
datapoint in payload_bytes_recv, but in 5 MB transfers has the
equivalent data point in payload_progress_recv.