Don't use entry guards in onion services
As in tor clients, using guards may be undesirable since we use a
single service process to represent multiple services.

Having guards enabled in onion services particularly doesn't work well
in tiny test networks such as in the tornettools CI. e.g. the service
can choose one or more of the guards as intro points, and then not be
able to find distinct guards to contact those intro points.