Log tor output on errors generating keys (#95)
While experimenting with a different configuration (compiling tor with asan), we hit this failure with no way to debug. We were already reporting the stderr when failing to generate keys for authorities, but not for other relays.

This changes both to redirect stderr to stdout and log the combined output on error.